requirements.txt
Django==4.0.3
python-socketio==5.5.2
